DEV: Updated wrapping for strange doctypes and output as Param
Numerous functions, at least in scipy include output in the
parameters list instead of as a returns list. and actually
don't have a Returns list. So these tools now will identify
output as an input parameter and will change it to a return
parameter.
Also now have a hierarchy for doc types that include multiple
input types (e.g. float and int, sequence of scalars, scalar
and tuple, etc.)