DEV: changed define_output_ports so that the_descrip eq short_descrip
The function that truncates parameter descriptions is included in
all of the methods for scraping 'Returns' parameters from the
docstrings.
I changed the truncated short descriptions (short_description) to
over write the variable 'the_description'. However, it currently
looks like the description parameter is not used except for in
the error message that occurs when output parameters are totally
absent from both the 'Parameters' and 'Returns' sections of the
docstring.