DEV: Removed description truncation from output scraping
Removed all references to the truncation function and
short_description from define_output_ports. If we need to add
them back in then they need to be inserted after 'the_type'
is assigned. Currently at lines: 541, 558 and 574. The length
of truncated descriptions also will need to be defined at the
beginning of the function