DEV: _ARRAY_SHAPE now detects dstring with (..., M), ndarray, array_like
Changed search function within parentheses to
\((([A-Za-z0-9]*[.]*, *)*[A-Za-z0-9]+,? *),?\) where
'\(' forces first detected character to be an opening parenthesis '('
the next parenthesis establishes search for a set, and the third
parenthesis starts definition of the first set.
[A-Za-z0-9]* searches for any alphanumeric character, and the *
establishes that those characters can be present zero or more times.
The addition of '[.]*' enables detection of zero or more dots (periods)
in series (e.g. '...', see numpy.linalg.svd() for official example).
the ',' establishes that detected string segments in this set must be
followed by a comma ',' and the ' *)' states that zero or more spaces
can be included after a preceeding tuple element.
'*' allows zero or more elements in this first set, and for zero or more
sets to be included.
The second '[A-Za-z0-9]+' requires that at least one (defined by the
'+') alphanumeric character set be included within the tuple in order
for the tuple to match this regular expression.
',?' indicates that the comma following last tuple element is optional.
' *' indicates that zero or more spaces can follow the last tuple
element, and ')' closes the tuple element search group.
As before, the ',?' indicates that an additional comma is optional,
and '\) *' forces a closing parenthesis, followed by zero or more spaces,
to be present in order for a match to occur.
Finally, '(array|ndarray|array_like)' the end of a matching string must
match either 'array', 'ndarray', or 'array_like' and all of these
constraints must be met in order for a string to match this RE.