DEV: added check for absent description to _truncate_description
Tom mentioned that there should be an if statement check to make sure
that a description is actually present for each parameter or return
variable being processed. I added a statment to make sure that the
original description being passed through actually contains at least one
string. If it doesn't the function will now return: ''